requirements.txt
numpy
pika
matplotlib
opencv-contrib-python
python-dotenv
requests
